+++ packages/components-react/accordion-react/src/index.ts
@@ -1,0 +1,11 @@
+import '@rijkshuisstijl-community/accordion-css';
+
+export {
+  Accordion,
+  type AccordionProps,
+  AccordionProvider,
+  type AccordionProviderProps,
+  AccordionSection,
+  type AccordionSectionProps,
+  type AccordionSectionProviderProps,
+} from './Accordion';