+++ packages/components-react/article-react/src/index.ts
@@ -1,0 +1,3 @@
+import '@rijkshuisstijl-community/article-css';
+
+export { Article, type ArticleProps } from './Article';