+++ packages/components-react/library-react/src/FileInput.tsx
@@ -4,9 +4,9 @@
  */
 
 import { ChangeEvent, PropsWithChildren, Ref, useEffect, useRef, useState } from 'react';
+import { Paragraph } from '.';
 import { Button, ButtonProps } from './Button';
 import { File } from './File';
-import { Paragraph } from './Paragraph';
 
 export interface FileInputProps extends Omit<ButtonProps, 'appearance'> {
   ref?: Ref<HTMLInputElement>;
@@ -16,8 +16,6 @@ export interface FileInputProps extends Omit<ButtonProps, 'appearance'> {
   allowedFileTypes: string;
   fileSizeErrorMessage: string;
   fileTypeErrorMessage: string;
-  // TODO: figure out why disabling is needed; works fine in editor but not in lint script for some reason
-  // eslint-disable-next-line no-unused-vars
   onValueChange?: (callbackFiles: File[]) => void;
   defaultFiles?: File[];
 }