+++ packages/components-react/library-react/src/PageHeader.tsx
@@ -3,4 +3,4 @@
  * Copyright (c) 2026 Community for NL Design System
  */
 
-export { PageHeader, type PageHeaderProps } from '@utrecht/component-library-react';
+export { PageHeader, type PageHeaderProps } from '@utrecht/page-header-react';