+++ packages/components-react/library-react/src/Root.test.tsx
@@ -1,0 +1,17 @@
+import '@testing-library/jest-dom/vitest';
+import { cleanup, render } from '@testing-library/react';
+import { afterEach, describe, expect, it } from 'vitest';
+import { Root } from '.';
+
+describe('Root', () => {
+  it('renders a visible element', () => {
+    const { container } = render(<Root Component="div" />);
+
+    const root = container.querySelector(':only-child');
+
+    expect(root).toBeInTheDocument();
+    expect(root).toBeVisible();
+  });
+});
+
+afterEach(() => cleanup());