+++ packages/storybook/src/components-react/link-button.stories.tsx
@@ -1,5 +1,4 @@
-import { LinkButton } from '@rijkshuisstijl-community/components-react';
-import { Icon } from '@rijkshuisstijl-community/icon-react';
+import { Icon, LinkButton } from '@rijkshuisstijl-community/components-react';
 import { Meta, StoryObj } from '@storybook/react-vite';
 
 const meta = {