+++ packages/storybook/src/types/drupal-attribute.d.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 declare module 'drupal-attribute' {
   export default class DrupalAttribute {
     constructor(data?: Record<string, any>);